+++ tests/phpunit/integration/AchieveTest.php
@@ -3,7 +3,7 @@
 namespace MediaWiki\Extension\AchievementBadges\Tests\Integration;
 
 use EchoNotificationMapper;
-use LogPage;
+use MediaWiki\Extension\AchievementBadges\Achievement;
 use MediaWiki\Extension\AchievementBadges\Constants;
 use MediaWikiIntegrationTestCase;
 use User;
@@ -40,15 +40,15 @@ class AchieveTest extends MediaWikiIntegrationTestCase {
 
 	private function assertEarnedAchievement( $num, $user, $key ) {
 		$dbr = wfGetDB( DB_REPLICA );
+		$query = Achievement::getQueryInfo( $dbr );
+		$query['conds'] = array_merge( $query['conds'], [
+			'log_action' => $key,
+			'log_actor' => $user->getActorId(),
+		] );
 		$actual = $dbr->selectRowCount(
-			'logging',
+			$query['tables'],
 			'*',
-			[
-				'log_type' => Constants::LOG_TYPE,
-				'log_action' => $key,
-				'log_actor' => $user->getActorId(),
-				$dbr->bitAnd( 'log_deleted', LogPage::DELETED_ACTION | LogPage::DELETED_USER ) . ' = 0 ',
-			]
+			$query['conds']
 		);
 		$this->assertSame( $num, $actual );
 	}